+++ src/main/kotlin/commands/Remove.kt
@@ -2,7 +2,7 @@ package com.neo.properties.commands
 
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.neo.properties.core.Command
-import com.neo.properties.errors.EnvironmentNotFound
+import com.neo.properties.error.EnvironmentNotFound
 import com.neo.properties.util.extension.json
 import com.neo.properties.util.extension.requireInstall
 import java.io.File
