+++ src/main/kotlin/commands/Save.kt
@@ -6,7 +6,7 @@ import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.mordant.terminal.YesNoPrompt
 import com.google.gson.Gson
 import com.neo.properties.core.Command
-import com.neo.properties.errors.TargetNotFound
+import com.neo.properties.error.TargetNotFound
 import com.neo.properties.util.extension.json
 import com.neo.properties.util.extension.readAsProperties
 import com.neo.properties.util.extension.requireInstall
