+++ src/main/kotlin/error/NotInstalledError.kt
@@ -1,4 +1,4 @@
-package errors
+package com.neo.properties.error
 
 import com.github.ajalt.clikt.core.CliktError
