+++ src/main/kotlin/util/extension/BaseCommand.kt
@@ -1,12 +1,11 @@
 package com.neo.properties.util.extension
 
 import com.github.ajalt.clikt.core.Abort
-import com.github.ajalt.clikt.core.CliktCommand
 import com.neo.properties.core.Command
+import com.neo.properties.error.NotInstalledError
 import com.neo.properties.util.Instructions
 import com.neo.properties.model.Config
 import com.neo.properties.util.Constants
-import errors.NotInstalledError
 import java.io.File
 
 fun Command.requireInstall(): Config {
